% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZanova.R
\name{EZanova}
\alias{EZanova}
\title{EZanova}
\usage{
EZanova(data, normal_detect = TRUE)
}
\arguments{
\item{data}{dataset (tibble or dataframe) to create model from}

\item{normal_detect}{does the function give a warning when data is not normally distributed? (Logical)}
}
\value{
aov object
}
\description{
Menu user-interfaced analysis of variance model creation
}
