% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZglm.R
\name{EZglm}
\alias{EZglm}
\title{EZglm}
\usage{
EZglm(data, type_detect = TRUE, normal_detect = TRUE, standardize_data = FALSE)
}
\arguments{
\item{data}{dataset (tibble or dataframe) to create model from}

\item{type_detect}{does the function automatically assign the Gaussian (linear) or Binomial (logistic) curve? (Logical)}

\item{normal_detect}{does the function give a warning when data is not normally distributed? (Logical)}

\item{standardize_data}{should the data be standardized for adjusted p and beta values? (Logical)}
}
\value{
glm object
}
\description{
Menu user-interfaced linear model creation
}
