% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZimport.R
\name{EZimport}
\alias{EZimport}
\title{EZimport}
\usage{
EZimport(data, name)
}
\arguments{
\item{data}{csv or excel file with data}

\item{name}{name of your EZimported file (character)}
}
\value{
EZimported file
}
\description{
Import large datasets and assign datatypes
}
