% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZrename.R
\name{EZrename}
\alias{EZrename}
\title{EZrename}
\usage{
EZrename(df)
}
\arguments{
\item{df}{your input dataframe}
}
\value{
df your dataframe with new column names
}
\description{
Menu interface for renaming columns in a large dataframe or tibble
}
