% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZsummary.R
\name{EZsummary}
\alias{EZsummary}
\title{EZsummary}
\usage{
EZsummary(model, column_menu = TRUE, round_digits = 4, styled = TRUE)
}
\arguments{
\item{model}{a model object (glm, aov, or manova)}

\item{column_menu}{Logical. If \code{TRUE}, displays a menu for user to manually choose summary stats. If \code{FALSE}, EZsummary will automatically display all stats}

\item{round_digits}{Number of digits to round numerical output to (numerical)}

\item{styled}{Logical. If \code{TRUE}, returns a styled gt table; if \code{FALSE}, r will return a tibble that is more piratical for future use}
}
\value{
either a styled table or a tibble. Also gives the users the chance to export it
}
\description{
Generates a clean, readable summary table from a glm() model which can be exported as CSV
}
